requirements.txt
Django==5.2.3
Pillow==11.0.0
django-tailwind==3.8.0
django-browser-reload==1.15.0
python-decouple==3.8
